+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Changelog
 - Prepared UI dialogs for exporting paper wallets
 - Prepared UI dialogs for importing paper wallets
 - Acceptance test for "Restore wallet with and without spending password" feature
+- Acceptance test for "Create wallet with spending password" feature
 
 ### Fixes